# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank of Blockchain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transactions
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsListResponse"
  /transactions/{id}:
    get:
      operationId: getTransactionById
      tags: [transactions]
      summary: Get transaction by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/validate:
    post:
      operationId: validateTransaction
      tags: [transactions]
      summary: Validate transaction (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /transactions/{id}/reject:
    post:
      operationId: rejectTransaction
      tags: [transactions]
      summary: Reject transaction (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /crypto-transfers:
    get:
      operationId: getCryptoTransfers
      tags: [crypto-transfers]
      summary: Get crypto transfers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Crypto transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoTransfersListResponse"
    post:
      operationId: createCryptoTransfer
      tags: [crypto-transfers]
      summary: Request a crypto transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCryptoTransferRequest"
      responses:
        "201":
          description: Crypto transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoTransfer"
  /crypto-transfers/{id}:
    get:
      operationId: getCryptoTransferById
      tags: [crypto-transfers]
      summary: Get crypto transfer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Crypto transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoTransfer"
  /crypto-transfers/{id}/validate:
    post:
      operationId: validateCryptoTransfer
      tags: [crypto-transfers]
      summary: Validate crypto transfer (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /crypto-transfers/{id}/reject:
    post:
      operationId: rejectCryptoTransfer
      tags: [crypto-transfers]
      summary: Reject crypto transfer (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [notifications]
      summary: Get notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create notification (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [bank_transfer, crypto_transfer, credit, debit]
        amount:
          type: number
        currency:
          type: string
          enum: [EUR, USD, BTC]
        status:
          type: string
          enum: [pending, processing, completed, rejected]
        description:
          type: string
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
      required:
        - id
        - userId
        - type
        - amount
        - currency
        - status
        - createdAt
    TransactionsListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transactions
        - total
        - page
        - limit
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
...
    CryptoTransfer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        cryptocurrency:
          type: string
          enum: [BTC, ETH, USDT]
        walletAddress:
          type: string
        network:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, rejected]
        rejectionReason:
          type: string
        txHash:
          type: string
        transactionId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
      required:
        - id
        - userId
        - amount
        - cryptocurrency
        - walletAddress
        - status
        - createdAt
    CryptoTransfersListResponse:
      type: object
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/CryptoTransfer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transfers
        - total
        - page
        - limit
    CreateCryptoTransferRequest:
      type: object
      properties:
        amount:
          type: number
        cryptocurrency:
          type: string
          enum: [BTC, ETH, USDT]
        walletAddress:
          type: string
        network:
          type: string
      required:
        - amount
        - cryptocurrency
        - walletAddress
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, error]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
    CreateNotificationRequest:
      type: object
      properties:
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, error]
        sendToAll:
          type: boolean
      required:
        - title
        - message
        - type
...